introduce a NUMAExecutor prototype
Combines TThreadExecutor, TProcessExecutor and libnuma to reduce the
number of remote access to memory by splitting the multithreaded
execution between the numa domains, concealing each partition to a
single domain. Assuming the proximity in memory of the consecutive
elements of the arguments, this will increase data locality in a simple
way.